Move syncplayer models into dedicated files
